+++ src/com/edu/realestate/dao/RealEstateDao.java
@@ -4,6 +4,6 @@ import com.edu.realestate.model.RealEstate;
 
 public interface RealEstateDao extends AbstractDao<RealEstate> {
 
-	int countRealEstates();
+	long countRealEstates();
 
 }
